invoice-extractor-saas/backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0

# Configuration and environment
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Redis and caching
redis==5.0.1

# AWS/Cloud storage
boto3==1.34.0

# AI/ML - Groq API (using aiohttp for HTTP requests)

# Image and PDF processing
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
pypdfium2==4.26.0
pdfplumber==0.9.0
pytesseract==0.3.10
PyMuPDF==1.23.8

# Data processing and export
pandas==2.1.4
openpyxl==3.1.2

# File handling
aiofiles==23.2.0

# Validation
email-validator==2.1.0

# GDPR compliance dependencies
cryptography>=41.0.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9

# Testing dependencies
reportlab==4.0.8